refactor(VideoRepository): remove interface
